REFACTOR: Grouped all the enums together
