+++ app/experience/components/ExperienceView.tsx
@@ -18,8 +18,8 @@ import RoleDatabaseKeys from "@/database/Roles/RoleDatabaseKeys";
 import rolesDatabase from "@/database/Roles/RoleDatabaseMap";
 import skillDatabaseMap from "@/database/Skills/SkillDatabaseMap";
 import SkillDatabaseKeys from "@/database/Skills/SkillDatabaseKeys";
-import ExperienceTypeEnum from "@/enums/ExperienceTypeEnum";
-import SkillTypesEnum from "@/enums/SkillTypesEnum";
+import ExperienceTypeEnum from "@/enums/Experience/ExperienceTypeEnum";
+import SkillTypesEnum from "@/enums/Skill/SkillTypesEnum";
 import useFuseSearch from "@/hooks/useFuseSearch";
 import FilterCategory from "@/interfaces/filters/FilterCategory";
 import MaterialGroupInterface from "@/interfaces/material/MaterialGroupInterface";
