+++ components/MaterialLists/MaterialTab.tsx
@@ -30,7 +30,7 @@ import projectDatabaseMap, {
 import rolesDatabase, {
   roleDatabaseKeys,
 } from "@/database/Roles/RoleDatabaseMap";
-import MaterialTypeEnum from "@/enums/MaterialTypeEnum";
+import MaterialTypeEnum from "@/enums/Material/MaterialTypeEnum";
 import MaterialGroupInterface from "@/interfaces/material/MaterialGroupInterface";
 import MaterialInterface from "@/database/Materials/MaterialInterface";
 import MaterialListProps from "@/interfaces/props/MaterialListProps";
