+++ enums/Project/ProjectCategoriesEnum.ts
