+++ enums/Skill/SkillCategoriesEnum.ts
